+++ test/PHPUnit/connecteur-type/glaneur/lib/GlaneurDocumentCreatorTest.php
@@ -4,21 +4,26 @@ require_once __DIR__ . "/../../../../../connecteur-type/glaneur/lib/GlaneurDocum
 class GlaneurDocumentCreatorTest extends PastellTestCase
 {
 
+    private const HELIOS_AUTOMATIQUE = 'helios-automatique';
+    private const IMPORTATION = 'importation';
 
     /**
      * @throws Exception
      */
     public function testCreateDocument()
     {
+        $notification = $this->getObjectInstancier()->getInstance(Notification::class);
+        $notification->add(1, 1, self::HELIOS_AUTOMATIQUE, self::IMPORTATION, false);
+
         $glaneurLocalDocumentCreator = $this->getObjectInstancier()->getInstance('GlaneurDocumentCreator');
 
         $glaneurLocalDocumentInfo = new GlaneurDocumentInfo(1);
-        $glaneurLocalDocumentInfo->nom_flux = 'helios-automatique';
+        $glaneurLocalDocumentInfo->nom_flux = self::HELIOS_AUTOMATIQUE;
         $glaneurLocalDocumentInfo->metadata = ['objet' => 'test_pes'];
         $glaneurLocalDocumentInfo->element_files_association = [
             'fichier_pes' => ['test.xml']
         ];
-        $glaneurLocalDocumentInfo->action_ok = 'importation';
+        $glaneurLocalDocumentInfo->action_ok = self::IMPORTATION;
         $glaneurLocalDocumentInfo->action_ko = 'erreur_import';
 
 
@@ -27,6 +32,11 @@ class GlaneurDocumentCreatorTest extends PastellTestCase
                 $glaneurLocalDocumentInfo,
                 __DIR__ . "/../fixtures/pes_exemple/"
             )
+        );
+        $journal_logs = $this->getJournal()->getAll(1, "", "", "", 0, 100);
+        $this->assertEquals(
+            "notification envoyée à eric@sigmalis.com",
+            $journal_logs[0]['message']
         );
     }
 
@@ -38,16 +48,14 @@ class GlaneurDocumentCreatorTest extends PastellTestCase
         $glaneurLocalDocumentCreator = $this->getObjectInstancier()->getInstance('GlaneurDocumentCreator');
 
         $glaneurLocalDocumentInfo = new GlaneurDocumentInfo(1);
-        $glaneurLocalDocumentInfo->nom_flux = 'helios-automatique';
+        $glaneurLocalDocumentInfo->nom_flux = self::HELIOS_AUTOMATIQUE;
         $glaneurLocalDocumentInfo->metadata = [];
         $glaneurLocalDocumentInfo->element_files_association = [
             'fichier_pes' => ['test.xml']
         ];
-        $glaneurLocalDocumentInfo->action_ok = 'importation';
+        $glaneurLocalDocumentInfo->action_ok = self::IMPORTATION;
         $glaneurLocalDocumentInfo->action_ko = 'erreur_import';
 
-
-        //$this->expectExceptionMessage("Le formulaire est incomplet : le champ «Objet» est obligatoire.");
         $this->assertNotEmpty(
             $glaneurLocalDocumentCreator->create(
                 $glaneurLocalDocumentInfo,